% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAGISurveyApp-package.R,
%   R/AAGISuveyApp-package.R
\docType{package}
\name{AAGISuveyApp-package}
\alias{AAGISuveyApp-package}
\title{AAGISuveyApp: AAGI Survey App}
\description{
A user-friendly interface to generate AAGI S&S feedback survey URLs.

A user-friendly interface to generate AAGI S&S feedback survey URLs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://aagi-aus.github.io/AAGISurveyApp/}
}

}
\author{
\strong{Maintainer}: Adam Sparks \email{adam.sparks@curtin.edu.au} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})

Authors:
\itemize{
  \item Adam Sparks \email{adam.sparks@curtin.edu.au} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})
}

Other contributors:
\itemize{
  \item Grains Research and Development Corporation (\href{https://ror.org/02xwr1996}{ROR}) (GRDC Project CUR2210-005OPX (AAGI-CU)) [funder, copyright holder]
}

}
\keyword{internal}
